quake: new Quake module that handles awesome restart gracefully
The previous RC file is moved as a module. We don't rely on the fact
that the next window to be managed after spawning a new terminal will
be this terminal. Instead, we rely on terminal name to know if we need
to manage them. For each console, we have a manage hook that tries to
identify if the window to manage is the console. In this case, it is
transformed and fitted as a Quake console.

Moreover, we clean things that look like QuakeConsole but that are not
because we already have one. For example, when switching from a
two-screen setup to a single-screen setup.