shifty: remove unwanted features
This is easier to make modifications if I don't have to ensure that
all features still work. The idea is that I will keep using
`awful.rules.rules` for most tasks except tag related tasks.

Also, fix the "startup" stuff to allow to specify a tag which will be
only used on startup.