xss-lock: don't suspend compositor on lock
Let's see if the confidentiality issue is still here. The problem with
suspending the compositor is that Firefox with EGL will freeze its
rendering. See https://bugzilla.mozilla.org/show_bug.cgi?id=1712665.

OK, the problem was that when screens are added/removed, i3lock is
unable to stay on top. I'll live with that for now.